core: Only mark successfully started checks as running
Previously we used to mark checks as running as soon as we tried to
launch them, no matter if we failed to send them to a worker or not.

With this patch, we're a lot more careful and only mark them as
running and update the statistics when we've managed to successfully
paw them off to one of our workers.

Original-patch-most-likely-by: Ton Voon <ton.voon@opsview.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>