lib/worker: Fix think/typo in loggin condition
Since we read() in a loop and we tag the filedescriptors with
O_NONBLOCK, we more or less expect to end up at the EAGAIN
state every time we reap a check that times out, gets killed
and gets reaped by the SIGCHLD reaper.

This patch just alters the conditions slightly so we never
log EAGAIN warnings from reading, or any errors from the final
read, instead of just dropping EAGAIN warnings from the final
read.

Signed-off-by: Andreas Ericsson <ae@op5.se>