Make core tap-tests pass once again
The original tests were written for Nagios 3. Most of them no longer
even compile, but they do produce an ungodly amount of warnings and
errors when one tries to.

We leave the obsoleted code in place, along with their buildrules,
in case someone wants to take a stab at fixing them, or at least as
a sort of map to what needs testing.

The important part is that 'make test' now works properly and can
be incorporated into automated buildsystems.

Signed-off-by: Andreas Ericsson <ae@op5.se>