core: Stop double-logging SIGHUP's
Previously we would log them twice. Once from the sighandler, stating
that we were going to shut down, and once from the main loop, stating
that we would restart. Often, they would follow each other immediately,
which was sort of confusing. With this patch, we only log the shutdown
thing if we actually are shutting down.

Signed-off-by: Andreas Ericsson <ae@op5.se>