cgi's: Refactor the whole config + status data reading somewhat
Most cgi's read everything in exactly the same way, but link to
different header and footer functions. By passing the header and
footer as arguments to the newly created cgi_init() function, we
can bypass that particular hurdle and unify how most the cgi's
process configuration and status data.

The ones we can't convert are the ones that sometimes generate images
and the mobile ones, but it's still a pretty neat win in code-size
reduction.

Signed-off-by: Andreas Ericsson <ae@op5.se>