cgis: Pass main config file to status data reading routines
It makes absolutely no sense what so ever to pass in the cgi.cfg file
location, as the xdata/xsddefault.c routines had to parse out the main
config file first and then parse that file to get the status_log
location. Since the cgi's always (without fail) know where the main
config file is by the time we read the status data, we might as well
pass the main config file directly and get rid of some 50 lines of
cumbersome code that just made it completely impossible to unify config
parsing in any way.

Signed-off-by: Andreas Ericsson <ae@op5.se>