Stop parsing config from xdata/xsddefault.c
We can easily make status_file a global variable instead and let it
be parsed from base/config.c and cgi/cgiutils.c. This patch lets us
drop quite a lot of almost-duplicated code, which is a Good Thing(tm).

It also takes us one step further in the war on the xdata directory.

Signed-off-by: Andreas Ericsson <ae@op5.se>