perfdata: Run host and service perfdata commands through workers
This means we parallelize the perfdata updating commands, so we no
longer stop doing *everything* just to run a command that ships off
perfdata somewhere.

Unfortunately, we can't do the same with the perfdata file processing
commands, since those generally move the file out of the way and we
have to re-open the file after they have successfully completed.

Writing to a temporary (but named) spoolfile which we later move to the
"real" perfdata file before we run the command would allow us to issue
a simple "rename()" call and then send the processing command to one of
our workers, but that's a larger patch, and the code works as-is so
I'm reluctant to fiddle with it more than necessary.

Signed-off-by: Andreas Ericsson <ae@op5.se>