Stop making xdata/ stuff almost-but-not-really configurable
Removing it means we can get rid of quite a lot of Makefile code,
as well as a ton of ./configure lines and the sometimes confusing
comments at the end of a ./configure run where we tell the user
that they've "chosen" some particular way of doing things.

While at it, we dig through the xdata/ files and simplify them
somewhat in order to prep them for the upcoming code-migration.
Worthwhile, since one can plainly see that comments, downtime and
perfdata code can just be moved without even changing buildrules.
Reflective changes of that insight are carried over to the cgi
Makefile, where the xcomment-cgi.o and xdowntime-cgi.o targets
are removed completely, as they weren't linked into aything
before anyway.

Signed-off-by: Andreas Ericsson <ae@op5.se>