core: Disallow zero retry- or check-intervals
If we *do* have zero retry or check-intervals and happen to disable
checks of the type that we have zero intervals for, we'll keep on
scheduling and rescheduling the same checks over and over to run at
the exact same timestamp, which will only cause us to peg the cpu
at 100% and log insane amounts of data to the debug log. Not so hot,
so let's avoid it whenever possible.

Signed-off-by: Andreas Ericsson <ae@op5.se>