core: Run all checks via scheduler
This removes the special-casing we used to have for checks that
were run because we needed to determine dependencies. The net result
is that we get rid of a few special cases and a whole lot of code.

While at it, we also log more carefully what's going on. I needed
that during development, and it may as well stay now that it's there.
If nothing else, it'll provide easy access to debugging material in
case we get bugreports on how checks and notifications are run now
that we've changed that.

The nice thing about this is that we don't have to spend a lot of time
fiddling with special cases, so bugs introduced by this patch are
likely fixed by removing special-case handling code (which tends to
be the least exercised and the most bug-prone anyways).

We also get sensible job-control, now that all checks pass through a
single channel instead of coming in from every which way.

Signed-off-by: Andreas Ericsson <ae@op5.se>