checks: Set service check state to something sensible on check timeout
When a check - host or service - fails to exit properly, it always
becomes CRITICAL. When an active host check times out, it goes into
UNKNOWN. When a service check times out, nothing is done to the state,
which led to my system telling me that a check that timed out was an
OK check.

This sets the state to whatever service_check_timeout_state has been
configured to, which is CRITICAL by default. That makes a lot more
sense and appears to be appropriate with regards to documentation,
common sense and similar sources of Great Wisdom.

Patch-fixed-by: Andreas Ericsson <ae@op5.se> (because Robin is lazy)
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>