core: Never reschedule services to run immediately
In case the service exited with an error that was *not* a timeout,
we would previously reschedule the check to be run immediately
instead of after a gracetime. Since check errors tend to occur
multiple times if run with the exact same parameters, the same
error could spam over and over and over in the logs and fill them
up completely.

This patch prevents that, using the currently prevailing check
interval to reschedule the service check.

Signed-off-by: Andreas Ericsson <ae@op5.se>