Added handler for SIGXFSZ
The SIGXFSZ signal will be received when the size of a file exceeds resource
limits. Previously this signal would cause core to terminate. The signal
handler now traps the signal, attempts to figure out which file my have
cause the signal to be received, and attempts to a log a message with that
information.

Unless the user has ridiculously low resource limits, this is most likely to
occur with the perfdata files. A message is logged no more frequently
than every 300 seconds.