core: Add 'core squeuestats' as a read-only query
The queryhandler API is seriously under-utilized. With this patch we
augment it somewhat to get some reasonably useful stats out of the
scheduling queue. Events are tracked from "remove_event()" and
"add_event()", which are the only two valid entry-points for the main
scheduling queue. The printed stats is simply a count of how events
scheduled, broken down per type.

Signed-off-by: Andreas Ericsson <ae@op5.se>