core: Be noisier when the event queue is abused or misused
Maintaining the event queue is our primary responsibility, or else
we'll fail to run checks, do maintenance and other similar things.
This patch adds checks that make sure we at least get to know if we
abuse the scheduling queue somehow (or if a module does).

Functionally, we now just refrain from tracking double-removes of
events, and also remove events that we try to add when they're
already in the queue.

Signed-off-by: Andreas Ericsson <ae@op5.se>