core: Bomb out early when we read unnamed objects
The previous strategy was to silently proceed without the unnamed
objects, but that's quite a bad idea. Funnily enough, we had
checks in place for the (arguably) less important objects, such
as dependencies and escalations, but not for hosts, contacts,
commands, services, timeperiods and the various group objects.

This should take care of tracker issue #479.

Signed-off-by: Andreas Ericsson <ae@op5.se>