xodtemplate: Avoid double-freeing faulty servicegroup members
Previously when we ran into a "members" directive of a servicegroup
where one or more services couldn't be found, we would free() the
input variable to xodtemplate_expand_services(), but without
resetting that variable in the object we were supposed to expand
members for, which caused a double free() when we later tried to
free() memory for the object in question.

This patch corrects that, removing the superfluous free() call.

Reported-by: Mark Frost <mfrost8@gmail.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>