xodtemplate: Free object names last in xodtemplate_free_memory()
When we run into double free()'s and debug the subsequent coredump,
the released memory is sometimes handed back to the kernel before
we can inspect it, or gets overwritten by antihack fortifications.
With this patch, we can be sure to find the names of the object
acting up when we inspect double-free coredumps, which often shortens
the investigation time by quite a large amount.

Signed-off-by: Andreas Ericsson <ae@op5.se>