downtime: Make downtime lookups a 16384 times more efficient
For a given value of "efficient" at least. Since we're using a fanout
table, we'll be using some memory in exchange for the fewer operations,
but at least lookup operations have decent practical performance now.

While at it, we sneak in the embryo of an API, which we can build on
later when refactoring the code into something sensible.

Signed-off-by: Andreas Ericsson <ae@op5.se>