downtime: Simplify the insertion sort somewhat
If scheduled_downtime_list is NULL, there's no need for us to enter
the iteration loop, so we can handle it the same we would as if we
were deferring downtime sorting and just prepend the downtime to
the (empty) scheduled_downtime_list.

We can do the same trick if the downtime_compar() function tells us
that the new downtime sorts before the first entry in the list.

Signed-off-by: Andreas Ericsson <ae@op5.se>