downtime: Handle error checking in downtime_add()
We're supposed to have all the information we need there, so we might
as well do all error checking there and get rid of the validation
routine in xdata/xdddefault.c, which is then reduced to setting
next_downtime_id = 1 in case nothing else sets it when we initialize.

Moving that small task to common/downtime.c as well means we can get
rid of the entire xdata/xdddefault.c, which is a very nice bonus.

Signed-off-by: Andreas Ericsson <ae@op5.se>