downtime: Ensure that scheduled_downtime_depth never drops below zero
If it does, subsequent downtimes may cause the depth to be bumped up
to zero, which will cause the notification machinery to not consider
the host or service to be in downtime at all, which in turn means
that notifications will be sent out while downtime is actually in
effect. It's unclear how one triggers this without modifying the
retention data file, but it has happened a few times in the wild,
and this rather simple patch efficiently protects against it.

Signed-off-by: Anton Lofgren <alofgren@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>