objectjson: Host details now use *member lists for hosts/services
Until now, the list of parent hosts and child hosts were generated
by iterating over all hosts to find the relationships. Now the lists
are generated based on the parent and child host structure members.
Similarly, the service list for a host was generated by iterating
over the list of all services and finding all services for the host.
Now the list uses the services structure member.