A couple hacks to make checks return output on RHEL/CentOS 5.
First, there appear to have been a race condition when workers
fork checks. It was at least theoretically possible that the
plugin would write to its stdout and/or stderr to the pipes
connecting to the worker before the worker registered the read
end of those pipes to the iobroker. This is now corrected
using a call back passed to the runcmd_open() function so the
iobroker_register() calls are made before the worker forks.

Second, it appears that checks are only able to return output
consistently on at least RHEL/CentOS 5 systems if iobroker
uses the select method rather than epoll or poll AND the check
sleeps for a short period after forking but before exec'ing the
plugin. For this reason a 100 ns sleep has been added to the
check at this point.