Really fixed bug #484 - Output from checks not getting back into Core
There was a race condition in the code that gathered the output from
checks where if the stderr was read first, stdout would never be read.
This commit fixes that condition.

This commit also revert the sleep introduced as part of commit
0d1895e38bfae197ce9dba295f1fec011461f66c. I'm still not sure why that
change worked, but it is no longer necessary as a result of the other
change in this commit.