Fixed bug 492: Nagios 4 fails to remove/add checks upon reload
The init script had been updated to use precached objects by default,
something which was not the behavior in Nagios Core 3 and is not
the expected behavior by most users. This commit reverts to the
previous behavior of not using precached objects by default. If
you want to use them, set the variable use_precached_objects to
"true" in /etc/sysconfig/nagios.

This commit also fixes the config check logic. Previously configcheck
was defaulted to false and then later checked to see whether it was
false and if so ran a configuration check. Now it is defaulted to
true and the configuration check runs only when set to true.