Added authorization checks for availability query.
The individual hosts and services that might become part of the
output in the availability query are now checked to see whether
the requesting user is authorized for the host or service.

In the event that a single host or service is requested, if the
requestor is not authorized for the host or service, the query
will fail silently and no data will be returned. This should not
be an issue because, if the requestor is not authorized for
a host or service, they would not know (from Nagios) that it
exists and therefore would not know to ask about it.