Fixed bug 507: Nagios 4.0.0 - Problem during log rotate (Stefano Ghelfi)
This is really a hack because the root cause is that some log rotations
are running early as a result the next log rotation is being scheduled
at the same time as the one that ran early.

This hack is being committed so that people will stop losing archives
because of the bug. A proper fix will be committed when the reason for
the early execution is understood.