Fixed bug #502: 4.0.0: Configuration -> Service Escalations = incomplete list
This was caused by trying to iterate over the list of service escalations
using the number of services as the limit. If you had more services than
service escalations, it would likely segfault. If you had fewer services
than service escalations, it would simply truncate the list of service
escalations. This commit causes the diplay of service escalations to use
the correct limit.