Macros: fixed issues with unallocated memory being freed.
At some point, most macro values were moved from allocated memory to
the stack, but the memory for the values was still being freed, which
was obviously causing problems. This commit moves the macro values
back to allocated memory, mostly thanks to the generous application of
strdup(). Eliminating the allocation/free cycle will be investigated
again in the future.

Also in this commit, macros for the new hourly value/importance host
and service attributes were temporarily removed. They were not completely
implemented and were causing problems.