Macros: Implemented importance (hourly value) macros
This commit implements macros for the hourly value attributes on hosts
and services. There are three new macros. HOSTIMPORTANCE is the
configured value of the host's hourly_value. HOSTANDSERVICESIMPORTANCE is
the configured value of the host's hourly_value plus the sum of the
configured hourly_value values for each of the host's services.
SERVICEIMPORTANCE is the configured value of a the service's hourly_value.

This commit also increases global code reuse by creating a separate
function, clear_custom_vars, to delete a list of custom variables.