Aliased hourly_value to importance and minimum_value to minimum_importance
This commit aliases the host and service hourly_value attribute to
importance and the contact attribute minimum_value to minimum_importance.
This is being done to reflect the more generic nature of these attributes.
It also deprecates the use of hourly_value and minimum_value. The names
of the structure members that store these values will remain unchanged
until Nagios Core 5.