Corrected calulation of host importance and importance defaults
This commit sets the default values for host and service importance and
contact minimum_importance to zero to agree with the documentation.

It also corrects the calculation of the host importance. Previously,
the value of each service was added to the host importance when the
service was added to the list of the host's services. Then when the
determination was being made whether to notify on host problems, the
sum of all the host's services' importance was added to the total
importance again, effectively doubling the importance of each of
the host's services. Now the host's stored importance is for only
the host itself, but it's services' importances are still used to
determine whether notifications should be sent.

Finally, the config CGI was modified to display the importance of the
host itself in one column and the importance of the host and it's services
in a second column.