Archive JSON: Last data update now set on warnings
This commit adds code to fetch a member of the JSON structure based on
a path to the member. In the event that a warning condition is
encountered because of passed parameters, this code is used to fetch
the result object so the last data update key can be updated.

This commit also initializes the value of the last data update so we
don't end up with last data update times thousands of years in the
future.