Fixed bug #573: Service checks returns (No output on stdout) stderr
This commits adds a check to the gather_output() function in lib/worker.c.
The check verifies that the reason for a failed read is NOT that the
socket cannot be read due to a temporary unavailablility (errno == EAGAIN).
If the socket is temporarily unavailable, the function simply tries the
read again.