xdata/status: Fix parameter passed to xsddefault_read_status_data()
Previously the name of the first parameter to xsddefault_read_status_data()
was cfgfile, but the value of the parameter was never used in the function.
Instead, the function used the global variable 'status_file'. In addition,
all functions that called xsddefault_read_status_data() and their ancestors
passed in the name of the main configuration file.

This commit fixes the semantics of the function so it is now passed the
name of the status file it actually reads.