Created initialization function for shared configuration variables
This commit adds an initialization function for configuration variables
shared between core and the CGIs. The function is necessary because
one of the variables (and possibly more in the future) was pointing to
data segment memory, but depending on the path of execution, could be
free()'d. Such a path was occurring in one of the test modules. The
initialization function strdups that variable to prevent the messiness
that occurs when freeing data segment memory.