Fix bug in check_window() macros
This commit corrects the logic in the check_window() macro so that
the macro returns the retry interval when the object (host or service)
is in a non-OK soft state and the check interval otherwise. Previously
it was returning the retry interval only when the object was in a soft
OK state meaning that checks for objects in OK states were happening
more often than necessary and checks for objects in non-OK states
weren't happening as often as they should.