Fix bug #596 (duplicates #590, #586): SIGSEGV crashes in CGIs.
If contact is not defined in authorized_for_all_hosts directive of cgi.cfg
and authorized_for_all_hosts is not *, most CGIs die with SIGSEGV.

is_contact_for_host_escalation() was referencing a NULL-initialized temporary
instead of the hostescalation argument. Correct the references, remove the
temporary, and add an argument check.

(Alexey Dvoryanchikov)