wproc: Make wproc_reap() behave as advertised.
Although only called when checking if an event should be run when job
control is enabled, this function did not behave as advertised.
Additionally waiting up to 3 seconds to reap 100 jobs may be excessive,
especially when fewer than 100 jobs are running. We'll wait up to one
millisecond for one job instead.