Check events: Don't estimate exec. time in auto-rescheduling.
Estimates of check execution time weren't too sophisticated or uniform
(we only have real measurments for host checks). By removing these
estimates the algirithm is simpler and the new schedule is smoother.

Also use cheduling_info.service_inter_check_delay/4 as a threshold for
determining if check events are too close in time and need to be
rescheduled. This removes hardcoded thresholds and makes the
rescheduling criteria dependent on the acutal configured checks.