runcmd: Continue to fork() unbalanced commands.
The worker layer & event/check scheduling aren't handling an error
return from runcmd_open() well. To get sensible handling of commands
with unbalanced quotes at the higher layers, continue on to fork a
child process that will exit with an error status and write a message
to stderr that will get picked up by the worker and sent to Core.
Currently, if we just return an error from the parent (worker) we can
end up in situations where the failing command will be repeatedly
rescheduled to run immediately.