runcmd: Fix build when we don't HAVE_SETENV.
A two character change (or move a '=' character...) to fix an
incredibly horrid mistake on my part that broke the build on Solaris,
OS X, and other systems that don't have setenv(). I didn't retest
without HAVE_SETENV after adding in the errror handling. I'm going to
frame this and put it on the wall next to the other smart things I've
done...