Map displays are now the full size of browser window
The map resizes as the browser window resizes. In addition the page
header on the map floats above the map and by default only shows a
single line, but it can be pinned.

Created a separate CSS file for the map directive and moved some other
sytle information out of D3.

Fixed an error with the display of icons on Firefox and Safari.