Report Time Period does not work if an @ character is in the timeperiod name
Fix for issue http://api.nagios.com/tps/viewtask.php?task_id=6464

A timeperiod of x@y is converted to x%40y when submitted to the website.
The cgi converts it back to x@y when processing the variables.
The Availability report then looks for a matching timeperiod, but does
so by converting x@y to x%40y, so the compair fails.