Network Status Map in Core 4.1 does not respect default settings in cgi.cfg
Fix for issues:
https://github.com/NagiosEnterprises/nagioscore/issues/37
https://github.com/NagiosEnterprises/nagioscore/issues/86

The new Status Map will utilize the `statusmap_background_image`
and the `default_statusmap_layout` options.

`default_statusmap_layout` will default to `6` if it is unset
on the new map. It can take a number in the range: 0-3,5-10.
Less than zero, greater than 10, or equal to 4 will instead use
a `6`. See the `sample-config/cgi.cfg` for the values.