nagios host URL parameter for NEW map doesn`t work - Network Map for All Hosts
Fix for issue https://github.com/NagiosEnterprises/nagioscore/issues/129

This is actually a partial fix. The requester wanted to limit the number
of hosts when the map is displayed from the hostgroup pages. In the new map
and the older CGI map, neither one limited the display to one or a few
hosts. But I did notice that the requested host was not being highlighted
in the new map. I fixed that by doubling the font size of the hostname,
and outlining it with red.