external command during reload doesnt work
Fix for issue https://github.com/NagiosEnterprises/nagioscore/issues/112

The function base/commands.c/command_input_handler() checks if the
`sigrestart` flag is set, indicating that it's in the process of doing a
reload. If it is set, preveiously the function would just discard the
command. Now, it will sleep for a while, then check again and process the
command once it's safe to do so.