Update daemon-init.in
If a stalled command pipe gets deleted, while there are processes trying/waiting to open it for writing, those processes get deadlocked in a blocking open() system call. In order to avoid such situations, or rather to give such processes a chance to die on a broken pipe, a pre-existing named pipe should shortly be _opened for reading_ (without actually reading from it) before discarding it. To avoid the chicken-egg-problem, the inode should be relabeled beforehand.

see http://unix.stackexchange.com/questions/335406/opening-named-pipe-blocks-forever-if-pipe-is-deleted-without-being-connected

A better solution would be, of course, to avoid the deletion of the pipe all together. I'm not quite sure why the that would be necessary anyway.