Vent command pipe before remove to avoid deadlocks
Removing a stalled command file, while there are processes trying/waiting to write into it, will deadlock those processes in a blocking open() system call. To allow such processes to die on a broken pipe, the pipe must be opened for reading without actually reading from it, which is what dd does here. To avoid a chicken-egg problem, the pipe is renamed beforehand. In order for the dd to not deadlock when there is no writing process, it is executed in the background in a subshell together with an empty echo to have at least one writing process.

see http://unix.stackexchange.com/questions/335406/opening-named-pipe-blocks-forever-if-pipe-is-deleted-without-being-connected