Replaced basic authentication with form authentication
Uses apache sessions and a login form for logging in and logging out.
Added a "Logout" link to the side menu.

In ./configure, there are now two cgi directory options. `cgiurl`
is the same as before, but now defaults to `cgibin` instead of
`cgi-bin`, used internally by the CGIs and web pages. The new one
is `cgi2url` (silly name, I know) which defaults to `cgi-bin`.
The two different url's allow for the sessions to work on the
normal cgi's, but allow basic authentication to work for scripts.
Quite a few people have scripts that have `http://.../nagios/cgi-bin`
hard-coded, and I don't want those to break.

To handle the two separate cgi url's, quite a few files had to have
(mostly) minor changes. It's possible (though unlikely) that I
missed something or broke something. If so, I'm sure issues will be
opened with all due haste in github.