Allow more flexible requirements for comments
Fix for issues #82 and #180

A section named `COMMAND COMMENTS` has been added to the end of `cgi.cfg`.

The configuration variable name is the command name, with `CMD_` replaced
with `CMT_` (for CoMmenT). So, for example, if you want to modify the
comment requirements for the `CMD_ACKNOWLEDGE_HOST_PROBLEM`, the name in
the `cgi.cfg` file would be `CMT_ACKNOWLEDGE_HOST_PROBLEM`.

The value part has two parameters. The first is a number that determines
if a comment is required for that command:

    A `0` (zero) means a comment is not allowed, and the `comment` fields
    on the form will not be displayed.

    A `1` (one) means that a comment is optional. The `comment` fields on
    the form will be displayed, but will not be marked as `required` (i.e.
    printed in red). If a comment is entered, it will be processed, but
    if no comment is entered, it will not be an error.

    A `2` means that a comment is required. The `comment` fields on the
    form will be displayed, and will be marked as `required` (i.e.
    printed in red). If a comment is entered, it will be processed, but
    if no comment is entered, it will be an error.

The second parameter is optional, and separated from the `required`
parameter by a comma. Everything after the comma is considered a default
comment for that command, and will be pre-loaded into the comment field
on the form.

Here are a couple of examples of things you can do:

CMT_ACKNOWLEDGE_HOST_PROBLEM=2,This problem is being looked into by [name]
This makes comments for host problem acknowledgements required (which
is the current behavior) but additionally, pre-loads the comment field
on the form with the value `This problem is being looked into by [name]`.

CMT_SCHEDULE_SVC_CHECK=1
This makes comments for rescheduling a service check optional. The
current behavior is that a comment is required.

CMT_SCHEDULE_HOST_DOWNTIME=0
This makes comments not allowed for scheduling downtime for a host. The
comment field on the form will not be displayed, where before, it was
displayed, and a comment was required.

CMT_DISABLE_SVC_CHECK=2
Normally, comments can not be entered when you disable active checks
for a service. This makes comments required, displays the comment fields
on the form, and gives an error if a comment has not been entered.

If no `COMMAND COMMENTS` configuration values are entered in `cgi.cfg`,
the defaults are the same as the current behavior.

Nagios core currently only has comments for host or service related
commands. But comment overrides for every command can be entered in
`cgi.cfg`. That means, if you put in `CMD_DISABLE_FLAP_DETECTION=2`,
disabling flap detection system-wide will have comment fields on the
form and will require a comment. But the comment will just be thrown
away. Some non-host/service commands may or may not allow and process
comments in the future.