new status for check dependencies
Enhancement for issue #229

There is a newo `SET SERVICE/HOST STATUS WHEN SERVICE CHECK SKIPPED`
section in `nagios.cfg`. There are four new settings that control
what status hosts and services should get under various circumstances.

Normally, if a check is not performed because of a failed dependency,
parent status, or host down status, the skipped check causes the
status of the host or service to stay the same. So if a host crashes,
the services on that host will never be checked until the host comes
back up. So the services could remain `STATE_OK`.

The new configuration settings will allow you to change that.

If a service check is skipped because of a failed dependency, the
`service_skip_check_dependency_status` setting is checked. If a
parent status causes a check to be skipped, the
`service_skip_check_parent_status` setting is checked. If the host
for a service is down, the `service_skip_check_host_down_status`
is checked. And if a host check is skipped because of a failed
dependency, the `host_skip_check_dependency_status` setting is
checked.

If the setting is `-1` (the default), nothing is done. Otherwise,
the status of the host or service is set to the value of the
setting. For services, value options are: 0 - STATE_OK,
1 - STATE_WARNING, 2 - STATE_CRITICAL, 3 - STATE_UNKNOWN. For
hosts, value options are: 0 - STATE_UP, 1 - STATE_DOWN,
2 - STATE_UNREACHABLE. One or more other settings may be added
in the future.

In most cases, if this is going to be used at all, the various
`service_skip_check_*_status` settings would probably be set to
`3` (STATE_UNKNOWN).

Conflicts:
	Changelog
	base/checks.c
	base/utils.c
	include/nagios.h