First commit for notification overrides
See issue #95

The idea here is to allow the overriding of notification flags on
a host or service level.

For example, say there is an accounting program on server "acctg",
and a service check that makes sure the software is running properly,
named "Accounting Working".

Bob is the manager of the accounting group. Like everyone else in
the accounting group, he gets notified for most services on the
"acctg" host only for critical and recovery (c,r) notifications.
However, he needs to be notified if the "Accounting Working"
service goes into a warning state.

In addition, Bob's boss Fred, who normally doesn't get notified
about any services in the "acctg" server, needs to be notified
if the "Accounting Working" is entering scheduled downtime.

You could define multiple groups, or multiple users, or multiple
service checks, or use some complicated escalations. Or you can
put a notification options override on the service (or host) check.

Look at the last line of the service definition. It says user bob
will get warning, critical, and recovery notifications, and fred
will get scheduled downtime notifications.

Notification option overrides can be put on either the "contacts"
entry in a host or service, or the "contact_groups" entry. Note
that a entry on the "contacts" entry will take precedence over
any groups the user is a part of, reguardless of whether or not
there are override flags on the group, the user, both, or neither.

    define contact{
        contact_name                 bob
        use                          generic-contact
        alias                        Bob from Accounting
        email                        bob@mycompany.co.zz
        service_notification_period  24x7
        host_notification_period     24x7
        service_notification_options c,r
        host_notification_options    d,u,r
    }

    define contact_group{
        contact_group_name           acctg-users
        alias                        Accounting Users
        members                      bob,sally,phil,larry,curly,moe
    }

    define service{
        use                          generic-service
        host_name                    acctg
        service_description          Accounting Working
        check_command                check_acct_ok!32!38
        contact_groups               admins,acctg-users
        contacts                     bob:wcr,fred:s
    }